More accurately match the HD shipyard graphics to its SD counterpart
Now properly overlays the screen edges to make it look like it's part of the UI frame.

Did color correcting on the doors and grey bits.
Moved the powerlines over to match the location of the SD version.

Moved the little light boxes around so they'd line up better overall.