HD: New artwork for various 3DO tiles
I did this to get rid of some questionable stock art that I found to not be legitimately licensed. (Pirated stock art from Adobe Stock that somehow found its way onto some "free" art sites for example)

It's still not perfect as the majority of the art I've replaced it all with has custom licenses that require attribution, but it's better than before.